Do not update lastfailed with --lf/--ff still failing
This helps the use case of working on a set of failed tests, where you
would the initially failed ones to be still run then.

I have ``--cache-readonly`` here to help with that already, but it is
certainly a hack.

I've thought about having ``--lfk`` (``--last-failed-keep``), but would
rather avoid a new option, and the behavior appears to be sane in
general.