TerminalReporter.ensure_newline: also wrap with anything written
This removes checking (and especially unsetting) of `currentfspath` for this.

Adds `TerminalReporter._enforce_newline_for_summary()` to enforce an
empty line before the first summary, which previously worked through
`currentfspath` not being `None`, but this method does not work anymore
when `currentfspath` would get set to `None` for when a newline gets
written via `_width_of_current_line` already now.
It also makes it more explicit in general.