assertion: rewrite: only catch EnvironmentError
This was changed unintentionally in 45c4a8fb3 (pytest 5.3.0), but only
EnvironmentErrors might have `errno`.
Since that is not really guaranteed and it is good to have more
information this uses the string representation of the exc in the trace
message.

(cherry picked from commit eac933acdede9aa623791f710199d9db2513a692)

Also changed for error with `open` on non-Windows implementation.