Remove "reraise" from CallInfo.from_call (#318)
It was only used to conditionally (not) reraise `KeyboardInterrupt`, but
only within tests (and not e.g. collection).

It also appears to not make sense to make this dependent on if `--pdb`
(`usepdb`) is used or not.
Instead, the "debugging" plugin hooks into `pytest_keyboard_interrupt`,
and handles being used from there.  It has to skip `Exit` (which
triggers `pytest_keyboard_interrupt` for historical reasons), and
`Session.Interrupted` (derived from `KeyboardInterrupt`) there then.

doc: fix "when" with CallInfo: "memocollect" was removed in e1674f60e.